Update memory_statistics to use config file
Since the library configuration is taken on the command line, it has to
be duplicated at all the call sites of the action. This has the action
get the library config from a config files so that workflows can use the
same config. This will also aid the action's use in the hub repos.