Add support for tagging files in memory tables (#27)
In memory statistic tables, we previously special case paths with 3rdparty in the path to tag them as third party dependencies. We now want to also tag first party deps with the source library. This will remove the special case and replace it with generic tagging that can be applied in the library's configuration. This will remain compatible with old config files, though libraries using the special case will need to update the config to explicitly specify the tag.